+++ src/chunkserver/chunkserver-common/global_shared_resources.h
@@ -32,3 +32,10 @@ inline std::mutex gDisksMutex;
 
 /// Container to reuse free condition variables (guarded by `gChunksMapMutex`)
 inline std::vector<std::unique_ptr<CondVarWithWaitCount>> gFreeCondVars;
+
+/// Active Disks scans in progress.
+/// Note: theoretically it would return a false positive if scans haven't
+/// started yet, but it's a _very_ unlikely situation.
+static std::atomic_int gScansInProgress(0);
+
+static std::atomic_bool gPerformFsync;
